Extract glob helpers that are needed in a CJS environment
